Support PLAY/STBY and STOP/STBY remote sequences
Previously, if STBY button on remote was pressed, device switched
between standby and active modes. There was no way to turn it on
from or turn it off by separate commands.

Now, fast (<200 ms) sequence of PLAY/STBY turns the device on (even
if it was on before) while sequence of STOP/STBY turns it off.

This can be used to guarantee the state when controlled from
e.g. smart home application.